# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Crypto Observer service API documentation.",
        "title": "Crypto Observer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/currency/add": {
            "post": {
                "description": "Добавление валюты в список валют для отслеживания.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "currency"
                ],
                "summary": "Добавление валюты",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Логин",
                        "name": "currencyID",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK - Currency added successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Currency ID is required",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/currency/price": {
            "post": {
                "description": "Получение цены валюты по ID",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "currency"
                ],
                "summary": "Получение цены валюты",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID валюты",
                        "name": "currencyID",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "timestamp",
                        "name": "timestamp",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Price of the currency",
                        "schema": {
                            "type": "number"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Currency ID is required",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/currency/remove": {
            "delete": {
                "description": "Удаление валюты в список валют для отслеживания.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "currency"
                ],
                "summary": "Удаление валюты",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID валюты",
                        "name": "currencyID",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK - Currency removed successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Currency ID is required",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}